Revert "Split translations from SuperCount and improve performance"
This reverts commit ec3d8a5412b0972d5e8b8fbb7e318b3b6e96c70a.

The changes depend on the presence of LiveTL. I tested the script in
incognito mode so that LiveTL would be disabled and it worked, but upon
explicitly disabling the extension it did not.

A future split is likely to occur after replicating LiveTL's faster
method of looking at past YouTube API requests to get message data.